move evaluate_position outside minimax and move copy_* functions to new class
