+++ backend/ai_django/app/game_classes/ai.py
@@ -16,6 +16,7 @@ if TYPE_CHECKING:
 import math
 import heapq
 from .game_logic import Logic
+from .copy_game_objects import CopyGameObjects
 from .constants import PIECE_TYPES
 from .move import Move
 PAWNS, KNIGHTS, BISHOPS, ROOKS, QUEENS, KINGS = PIECE_TYPES
@@ -57,6 +58,34 @@ class TreeNode:
 
 class ChessAI:
     @staticmethod
+    def evaluate_position(
+        player: 'PlayerType',
+        opponent: 'PlayerType',
+        material: 'MaterialType',
+        game_status: 'GameStatusType'
+    ) -> float:
+        if game_status.winner is not None:
+            return (-1) ** (game_status.winner == 'black') * math.inf
+
+        if game_status.game_result == 'draw':
+            return 0
+
+        num_legal_moves = [player.num_legal_moves, opponent.num_legal_moves]
+        mat = [0, 0]
+        for i, p in enumerate((player, opponent)):
+            mat[i] += (
+                material[p.colour][PAWNS]
+                + material[p.colour][KNIGHTS] * 3
+                + material[p.colour][BISHOPS] * 3
+                + material[p.colour][ROOKS] * 5
+                + material[p.colour][QUEENS] * 9
+            )
+
+        white_player = 1 - int(player.colour == 'white')
+        black_player = 1 - white_player
+        return mat[white_player] - mat[black_player] + (num_legal_moves[white_player] - num_legal_moves[black_player]) / 50
+
+    @staticmethod
     def calculate_deep_moves(
         currentTree: 'TreeNode',
         depth: int,
@@ -69,37 +98,8 @@ class ChessAI:
         move_history = currentTree.game_state_after_move.move_history
         game_status = currentTree.game_state_after_move.game_status
 
-
-        # TODO: move this function outside
-        def evaluate_position(
-            player: 'PlayerType',
-            opponent: 'PlayerType',
-            material: 'MaterialType',
-            game_status: 'GameStatusType'
-        ) -> float:
-            if game_status.winner is not None:
-                return (-1) ** (game_status.winner == 'black') * math.inf
-
-            if game_status.game_result == 'draw':
-                return 0
-
-            num_legal_moves = [player.num_legal_moves, opponent.num_legal_moves]
-            mat = [0, 0]
-            for i, p in enumerate((player, opponent)):
-                mat[i] += (
-                    material[p.colour][PAWNS]
-                    + material[p.colour][KNIGHTS] * 3
-                    + material[p.colour][BISHOPS] * 3
-                    + material[p.colour][ROOKS] * 5
-                    + material[p.colour][QUEENS] * 9
-                )
-
-            white_player = 1 - int(player.colour == 'white')
-            black_player = 1 - white_player
-            return mat[white_player] - mat[black_player] + (num_legal_moves[white_player] - num_legal_moves[black_player]) / 50
-
         if depth == 0:
-            currentTree.eval = evaluate_position(player, opponent, material, game_status)
+            currentTree.eval = ChessAI.evaluate_position(player, opponent, material, game_status)
             return currentTree
 
         curr_eval = (-1) ** (player.colour == 'white') * math.inf
@@ -111,12 +111,12 @@ class ChessAI:
                     from_loc = move.from_loc
                     to_loc = move.to_loc
 
-                    temp_player = Logic.copy_player(player)
-                    temp_opponent = Logic.copy_player(opponent)
+                    temp_player = CopyGameObjects.copy_player(player)
+                    temp_opponent = CopyGameObjects.copy_player(opponent)
                     temp_board = Logic.get_board_from_pieces(temp_player.pieces, temp_opponent.pieces)
                     temp_board_str = get_board_string(temp_board)
-                    temp_material = Logic.copy_material(material)
-                    temp_move_history = Logic.copy_move_history(move_history)
+                    temp_material = CopyGameObjects.copy_material(material)
+                    temp_move_history = CopyGameObjects.copy_move_history(move_history)
 
                     game_status = Logic.make_move(
                         temp_board,
@@ -136,7 +136,7 @@ class ChessAI:
                         game_status
                     )
                     child_tree_node = TreeNode(child_game_state)
-                    child_tree_node.eval = evaluate_position(temp_player, temp_opponent, temp_material, game_status)
+                    child_tree_node.eval = ChessAI.evaluate_position(temp_player, temp_opponent, temp_material, game_status)
                     legal_moves.append(child_tree_node)
         legal_moves.sort(key=lambda state: state.eval, reverse=player.colour == 'white')
         max_or_min = max if player.colour == 'white' else min
